requirements.txt
pandas==2.2.3
pdf2image==1.17.0
pdfplumber==0.11.4
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.3.3
python-docx==1.1.2
python-pptx==1.0.2
